+++ soynlp/tokenizer/__init__.py
@@ -1,5 +1,5 @@
-from ._tokenizer import LTokenizer
-from ._tokenizer import MaxScoreTokenizer
-from ._tokenizer import RegexTokenizer
-from ._noun_tokenizer import NounLMatchTokenizer
-from ._noun_tokenizer import NounMatchTokenizer
+from .tokenizer import LTokenizer
+from .tokenizer import MaxScoreTokenizer
+from .tokenizer import RegexTokenizer
+from .noun_tokenizer import NounLMatchTokenizer
+from .noun_tokenizer import NounMatchTokenizer
